Let early DevVenus upload
Currently replaces mars but should be a good start for a better Venus,
And hopefully a public release. This is in the mars folder because any
file renames or movements causes micdoodle8 core to crash. This replaces
the original mars and if you try to land on the mars you crash.